SN-48 removed deprecated code, finished some TODO marks, reorganized code for better readability; SN-47 removed deprecated code; SN-51 created fresh readme with better structure which now should include new endpoints; SN-25 new endpoints in router.go for new methods

+++ backend/pkg/handler/groupMemberHandler.go
@@ -21,34 +21,35 @@ func NewGroupMemberHandler(repo *repository.GroupMemberRepository) *GroupMemberH
 // and the ID of the user. It inserts a new row into the group_members table in the database,
 // which represents the user being a member of the group. If the operation is successful,
 // it returns nil. If there is an error, it returns the error.
-func (h *GroupMemberHandler) AddMemberHandler(w http.ResponseWriter, r *http.Request) {
-    vars := mux.Vars(r)
-    groupId, ok := vars["groupId"]
-    if !ok {
-        http.Error(w, "Missing group ID", http.StatusBadRequest)
-        return
-    }
-    intGroupId, err := strconv.Atoi(groupId); if err != nil {
-        http.Error(w, "Failed to convert groupid string to int: " + err.Error(), http.StatusBadRequest)
-        return
-    }
+// // DEPRECATED: Adding a member is now done in ApproveGroupMembershipHandler and AcceptGroupInvitationHandler
+// func (h *GroupMemberHandler) AddMemberHandler(w http.ResponseWriter, r *http.Request) {
+//     vars := mux.Vars(r)
+//     groupId, ok := vars["groupId"]
+//     if !ok {
+//         http.Error(w, "Missing group ID", http.StatusBadRequest)
+//         return
+//     }
+//     intGroupId, err := strconv.Atoi(groupId); if err != nil {
+//         http.Error(w, "Failed to convert groupid string to int: " + err.Error(), http.StatusBadRequest)
+//         return
+//     }
 
-    userId, ok := vars["userId"]
-    if !ok {
-        http.Error(w, "Missing user ID", http.StatusBadRequest)
-        return
-    }
-    intUserId, err := strconv.Atoi(userId); if err != nil {
-        http.Error(w, "Failed to convert userid string to int: " + err.Error(), http.StatusBadRequest)
-        return
-    }
+//     userId, ok := vars["userId"]
+//     if !ok {
+//         http.Error(w, "Missing user ID", http.StatusBadRequest)
+//         return
+//     }
+//     intUserId, err := strconv.Atoi(userId); if err != nil {
+//         http.Error(w, "Failed to convert userid string to int: " + err.Error(), http.StatusBadRequest)
+//         return
+//     }
 
-    err = h.repo.AddMemberToGroup(intGroupId, intUserId); if err != nil {
-        http.Error(w, "Failed to add member to group: "+err.Error(), http.StatusInternalServerError)
-        return
-    }
-    w.WriteHeader(http.StatusOK)
-}
+//     err = h.repo.AddMemberToGroup(intGroupId, intUserId); if err != nil {
+//         http.Error(w, "Failed to add member to group: "+err.Error(), http.StatusInternalServerError)
+//         return
+//     }
+//     w.WriteHeader(http.StatusOK)
+// }
 
 // RemoveMemberFromGroup removes a user from a group. It takes two parameters: the ID of the group
 // and the ID of the user. It deletes the row from the group_members table in the database that
@@ -75,7 +76,17 @@ func (h *GroupMemberHandler) RemoveMemberHandler(w http.ResponseWriter, r *http.
         http.Error(w, "Failed to convert userid string to int: " + err.Error(), http.StatusBadRequest)
         return
     }
-    // TODO: implement logic to check if the user trying to remove the member is the owner of the group
+    requestingUserId := r.Context().Value("AuthUserID").(int)
+    // logic to check if the user trying to remove the member is the owner of the group
+    isAuthenticated, err := h.repo.IsUserGroupOwner(requestingUserId, intGroupId)
+    if err != nil {
+        http.Error(w, "Failed to check if user is group owner: "+err.Error(), http.StatusInternalServerError)
+        return
+    }
+    if !isAuthenticated {
+        http.Error(w, "User requesting the removal is not the group owner", http.StatusUnauthorized)
+        return
+    }
 
     err = h.repo.RemoveMemberFromGroup(intGroupId, intUserId)
     if err != nil {
@@ -105,6 +116,8 @@ func (h *GroupMemberHandler) RequestGroupMembershipHandler(w http.ResponseWriter
         http.Error(w, err.Error(), http.StatusInternalServerError)
         return
     }
+    // TODO: this should notify the group admin that a user has requested to join the group
+    // TODO: this should notify the user that their request was sent and is pending
 
     w.WriteHeader(http.StatusCreated)
 }
@@ -123,11 +136,16 @@ func (h *GroupMemberHandler) ApproveGroupMembershipHandler(w http.ResponseWriter
         return
     }
     // get groupId from the request invitation that was accepted
-    //groupInvitation, err := h.repo.GetGroupInvitationByID(id)
-
+    groupInvitation, err := h.repo.GetGroupInvitationByID(id); if err != nil {
+        http.Error(w, err.Error(), http.StatusInternalServerError)
+        return
+    }
     // this should add player to group
-    //err = h.repo.AddMemberToGroup(groupInvitation.GroupId, groupInvitation.JoinUserId)
-
+    err = h.repo.AddMemberToGroup(groupInvitation.GroupId, groupInvitation.JoinUserId)
+    if err != nil {
+        http.Error(w, "Error adding member to the group: " +err.Error(), http.StatusInternalServerError)
+        return
+    }
     // TODO: this should notify the user that their request was approved
     // TODO: this should delete the invitation from the database
     w.WriteHeader(http.StatusOK)
@@ -151,16 +169,6 @@ func (h *InvitationHandler) DeclineGroupMembershipHandler(w http.ResponseWriter,
     w.WriteHeader(http.StatusOK)
 }
 
-// GetAllInvitationsHandler gets all invitations
-func (h *InvitationHandler) GetAllGroupInvitationsHandler(w http.ResponseWriter, r *http.Request) {
-    invitations, err := h.repo.GetAllGroupInvitations()
-    if err != nil {
-        http.Error(w, err.Error(), http.StatusInternalServerError)
-        return
-    }
-    json.NewEncoder(w).Encode(invitations)
-}
-
 // InviteMemberHandler sends an invitation to a user to join a group.
 // It creates an invitation in the database that can be accepted or declined by the user.
 func (h *InvitationHandler) InviteGroupMemberHandler(w http.ResponseWriter, r *http.Request) {
@@ -177,20 +185,6 @@ func (h *InvitationHandler) InviteGroupMemberHandler(w http.ResponseWriter, r *h
     }
     // TODO: this should notify the user that they have been invited to join a group
     w.WriteHeader(http.StatusCreated)
-    //json.NewEncoder(w).Encode(newInvitation)
-}
-
-// GetInvitationByIDHandler gets an invitation by ID
-// TODO: refactor to use userid from cookie to get all group invitations for the user (pending)
-func (h *InvitationHandler) GetGroupInvitationByIDHandler(w http.ResponseWriter, r *http.Request) {
-    vars := mux.Vars(r)
-    id := vars["id"]
-    invitation, err := h.repo.GetGroupInvitationByID(id)
-    if err != nil {
-        http.Error(w, "Failed to get invitation: "+err.Error(), http.StatusInternalServerError)
-        return
-    }
-    json.NewEncoder(w).Encode(invitation)
 }
 
 // AcceptGroupInvitationHandler allows a user to accept an invitation to join a group.
@@ -202,8 +196,17 @@ func (h *GroupMemberHandler) AcceptGroupInvitationHandler(w http.ResponseWriter,
         http.Error(w, "Failed to accept invitation: "+err.Error(), http.StatusInternalServerError)
         return
     }
-    // TODO: this should add player to group
-
+    // get groupId from the request invitation that was accepted
+    groupInvitation, err := h.repo.GetGroupInvitationByID(id); if err != nil {
+        http.Error(w, err.Error(), http.StatusInternalServerError)
+        return
+    }
+    //  add player to group
+    err = h.repo.AddMemberToGroup(groupInvitation.GroupId, groupInvitation.JoinUserId)
+    if err != nil {
+        http.Error(w, "Error adding member to the group: " +err.Error(), http.StatusInternalServerError)
+        return
+    }
     // FUTURE TODO: this notify the group list of the new member
     w.WriteHeader(http.StatusOK)
 }
@@ -220,4 +223,27 @@ func (h *InvitationHandler) DeclineGroupInvitationHandler(w http.ResponseWriter,
     // TODO: this should notify the invitation sender that the invitation was declined
     // it also then should delete the invitation from the database
     w.WriteHeader(http.StatusOK)
+}
+
+// GetInvitationByIDHandler gets an invitation by ID
+// TODO: refactor to use userid from cookie to get all group invitations for the user (pending)
+func (h *InvitationHandler) GetGroupInvitationByIDHandler(w http.ResponseWriter, r *http.Request) {
+    vars := mux.Vars(r)
+    id := vars["id"]
+    invitation, err := h.repo.GetGroupInvitationByID(id)
+    if err != nil {
+        http.Error(w, "Failed to get invitation: "+err.Error(), http.StatusInternalServerError)
+        return
+    }
+    json.NewEncoder(w).Encode(invitation)
+}
+
+// GetAllInvitationsHandler gets all invitations (ok)
+func (h *InvitationHandler) GetAllGroupInvitationsHandler(w http.ResponseWriter, r *http.Request) {
+    invitations, err := h.repo.GetAllGroupInvitations()
+    if err != nil {
+        http.Error(w, err.Error(), http.StatusInternalServerError)
+        return
+    }
+    json.NewEncoder(w).Encode(invitations)
 }
